added the left part of home page
