+++ src/components/core/atoms/Icon/Icon.js
@@ -4,8 +4,10 @@ import cross from '../../../../misc/images/cross.png'
 import v from '../../../../misc/images/v.png'
 import fileIcon from '../../../../misc/images/fileIcon.png'
 import profIcon from '../../../../misc/images/profIcon.png'
+import logo from '../../../../misc/images/logo.png'
 
-const Icon = ({ icon, heightParam, widthParam, degree, ...props }) => {
+const Icon = ({ icon, heightParam, widthParam, degree, additionalStyles, ...props }) => {
+  const [styles] = useState(additionalStyles)
   const [height] = useState(heightParam)
   const [width] = useState(widthParam)
   const [rotateDegree, setRotateDegree] = useState(degree)
@@ -13,7 +15,8 @@ const Icon = ({ icon, heightParam, widthParam, degree, ...props }) => {
     cross: cross,
     v: v,
     file: fileIcon,
-    person: profIcon
+    person: profIcon,
+    logo: logo
   })
 
   useEffect(() => {
@@ -22,6 +25,7 @@ const Icon = ({ icon, heightParam, widthParam, degree, ...props }) => {
 
   return(
     <IconStyled
+      styles = { styles }
       rotateDegree = { rotateDegree }
       height = { height }
       width = { width }
