+++ src/components/styled/atoms/iconStyled.js
@@ -5,6 +5,7 @@ height: ${ props => props.height };
 width: ${ props => props.width};
 transform: rotate(${ props => props.rotateDegree }deg);
 transition: 0.5s ease all;
+${ props => props.styles }
 `
 
 export default IconStyled
