+++ src/components/styled/pages/addFilePageStyled.js
@@ -2,13 +2,12 @@ import styled from 'styled-components'
 import colorScheme from '../../../misc/colorScheme'
 
 const HomePageStyled = styled.div`
-width: 100%;
 height: 100%;
+width: 100%;
 background-color: ${ colorScheme.marigold };
 display: grid;
-grid-template-columns: 50% 50%;
-grid-template-rows: 50% 50%;
-grid-auto-flow: column;
+grid-template-columns: 34% 33% 33%;
+grid-template-rows: 100%;
 `
 
 export default HomePageStyled
